fix seccomp type patch and compilable filters
