+++ complete_filter/bifrostFilterSeq.c
@@ -2,7 +2,8 @@
 #include <linux/bpf.h>
 #include <linux/unistd.h>
 #include <linux/ptrace.h>
-#include <bpf_helpers.h>
+#include <linux/types.h>
+#include <bpf/bpf_helpers.h>
 #include <linux/errno.h>
 #include <sys/types.h>
 #include "tracer_id.h"
@@ -17,15 +18,70 @@ struct node_edge {
     uint16_t syscall_id;;
 };
 
-DEFINE_BPF_MAP(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
-DEFINE_BPF_MAP(thread_node_id, HASH, uint64_t, uint16_t, 512);
-DEFINE_BPF_MAP(seq_tree_map, HASH, struct node_edge, uint16_t, 1278);
-DEFINE_BPF_MAP(init_var, HASH, int, int, 1);
+#define DEFINE_BPF_MAP_NO_ACCESSORS(the_map, TYPE, KeyType, ValueType, num_entries) \
+struct {                                                                         \
+        __uint(type, BPF_MAP_TYPE_##TYPE);                                       \
+        __type(key, KeyType);                                                    \
+        __type(value, ValueType);                                                \
+        __uint(max_entries, num_entries);                                        \
+        } the_map SEC(".maps");                                                  \
+
+
+#define DEFINE_BPF_MAP_N(the_map, TYPE, KeyType, ValueType, num_entries)         \
+        DEFINE_BPF_MAP_NO_ACCESSORS(the_map, TYPE, KeyType, ValueType, num_entries)     \
+static inline __always_inline __unused ValueType* bpf_##the_map##_lookup_elem(               \
+            const KeyType* k) {                                                                  \
+        return bpf_map_lookup_elem(&the_map, k);                                          \
+    };                                                                                           \
+                                                                                                 \
+    static inline __always_inline __unused int bpf_##the_map##_update_elem(                      \
+            const KeyType* k, const ValueType* v, unsigned long long flags) {                    \
+        return bpf_map_update_elem(&the_map, k, v, flags);                                \
+    };                                                                                           \
+                                                                                                 \
+    static inline __always_inline __unused int bpf_##the_map##_delete_elem(const KeyType* k) {   \
+        return bpf_map_delete_elem(&the_map, k);                                          \
+    };
+
+
+#define DEFINE_BPF_MAP_NO_ACCESSORS_F(the_map, TYPE, KeyType, ValueType, num_entries, flag) \
+struct {                                                                        \
+        __uint(type, BPF_MAP_TYPE_##TYPE);                                      \
+        __type(key, KeyType);                                                   \
+        __type(value, ValueType);                                               \
+        __uint(max_entries, num_entries);                                       \
+        __uint(map_flags, flag);                                                \
+        } the_map SEC(".maps");                                                 \
+
+#define DEFINE_BPF_MAP_F(the_map, TYPE, KeyType, ValueType, num_entries, flag)         \
+        DEFINE_BPF_MAP_NO_ACCESSORS_F(the_map, TYPE, KeyType, ValueType, num_entries, flag) \
+        static inline __always_inline __unused ValueType* bpf_##the_map##_lookup_elem(               \
+            const KeyType* k) {                                                                  \
+        return bpf_map_lookup_elem(&the_map, k);                                          \
+    };                                                                                           \
+                                                                                                 \
+    static inline __always_inline __unused int bpf_##the_map##_update_elem(                      \
+            const KeyType* k, const ValueType* v, unsigned long long flags) {                    \
+        return bpf_map_update_elem(&the_map, k, v, flags);                                \
+    };                                                                                           \
+                                                                                                 \
+    static inline __always_inline __unused int bpf_##the_map##_delete_elem(const KeyType* k) {   \
+        return bpf_map_delete_elem(&the_map, k);                                          \
+    };
+
+
+DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
+DEFINE_BPF_MAP_N(thread_node_id, HASH, uint64_t, uint16_t, 512);
+DEFINE_BPF_MAP_N(seq_tree_map, HASH, struct node_edge, uint16_t, 9051);
+DEFINE_BPF_MAP_N(init_var, HASH, int, int, 1);
+
+/*
 #define bpf_printk(fmt, ...)                                   \
 ({                                                             \
     char ____fmt[] = fmt;                                      \
     bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
 })
+*/
 
 void __always_inline init_seq_tree_map() {
     struct node_edge ne;
@@ -2591,6 +2647,8 @@ void __always_inline init_seq_tree_map() {
 
 SEC("seccomp")
 int __always_inline filter_seq(struct seccomp_data *ctx) {
+
+    (void)ctx;
     int ret = SECCOMP_RET_ALLOW;
     uint64_t pid_tgid = bpf_get_current_pid_tgid();
     struct syscall_info *info = bpf_syscall_info_map_lookup_elem(&pid_tgid);
