+++ complete_filter/fstat.c
@@ -2,10 +2,63 @@
 #include <linux/bpf.h>
 #include <linux/unistd.h>
 #include <linux/ptrace.h>
-#include <bpf_helpers.h>
+#include <linux/types.h>
+#include <bpf/bpf_helpers.h>
 #include <linux/errno.h>
 #include <sys/types.h>
 #include "tracer_id.h"
+
+#define DEFINE_BPF_MAP_NO_ACCESSORS(the_map, TYPE, KeyType, ValueType, num_entries) \
+struct {                                                                            \
+    __uint(type, BPF_MAP_TYPE_##TYPE);                                              \
+    __type(key, KeyType);                                                           \
+    __type(value, ValueType);                                                       \
+    __uint(max_entries, num_entries);                                               \
+} the_map SEC(".maps");
+
+
+#define DEFINE_BPF_MAP_N(the_map, TYPE, KeyType, ValueType, num_entries)              \
+    DEFINE_BPF_MAP_NO_ACCESSORS(the_map, TYPE, KeyType, ValueType, num_entries)       \
+static inline __always_inline __unused ValueType* bpf_##the_map##_lookup_elem(      \
+    const KeyType* k) {                                                             \
+    return bpf_map_lookup_elem(&the_map, k);                                        \
+};                                                                                  \
+                                                                                    \
+static inline __always_inline __unused int bpf_##the_map##_update_elem(              \
+    const KeyType* k, const ValueType* v, unsigned long long flags) {               \
+    return bpf_map_update_elem(&the_map, k, v, flags);                              \
+};                                                                                  \
+                                                                                    \
+static inline __always_inline __unused int bpf_##the_map##_delete_elem(const KeyType* k) { \
+    return bpf_map_delete_elem(&the_map, k);                                        \
+};
+
+
+#define DEFINE_BPF_MAP_NO_ACCESSORS_F(the_map, TYPE, KeyType, ValueType, num_entries, flag) \
+struct {                                                                            \
+    __uint(type, BPF_MAP_TYPE_##TYPE);                                              \
+    __type(key, KeyType);                                                           \
+    __type(value, ValueType);                                                       \
+    __uint(max_entries, num_entries);                                               \
+    __uint(map_flags, flag);                                                        \
+} the_map SEC(".maps");
+
+
+#define DEFINE_BPF_MAP_F(the_map, TYPE, KeyType, ValueType, num_entries, flag)        \
+    DEFINE_BPF_MAP_NO_ACCESSORS_F(the_map, TYPE, KeyType, ValueType, num_entries, flag) \
+static inline __always_inline __unused ValueType* bpf_##the_map##_lookup_elem(      \
+    const KeyType* k) {                                                             \
+    return bpf_map_lookup_elem(&the_map, k);                                        \
+};                                                                                  \
+                                                                                    \
+static inline __always_inline __unused int bpf_##the_map##_update_elem(              \
+    const KeyType* k, const ValueType* v, unsigned long long flags) {               \
+    return bpf_map_update_elem(&the_map, k, v, flags);                              \
+};                                                                                  \
+                                                                                    \
+static inline __always_inline __unused int bpf_##the_map##_delete_elem(const KeyType* k) { \
+    return bpf_map_delete_elem(&the_map, k);                                        \
+};
 
 struct syscall_info {
     uint8_t id;
@@ -35,13 +88,14 @@ struct stat {
     uint32_t __unused5; //const
 };
 
-DEFINE_BPF_MAP(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
+DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
+/*
 #define bpf_printk(fmt, ...)                                   \
 ({                                                             \
     char ____fmt[] = fmt;                                      \
     bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
 })
-
+*/
 SEC("seccomp")
 int __always_inline filter_fstat(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
@@ -53,9 +107,9 @@ int __always_inline filter_fstat(struct seccomp_data *ctx) {
 
     //arg ptr[out, stat] ptr 0xf8ffa0 8
     //arg stat stat 0xf8ffa0 128
-    struct stat v58;
-    if (bpf_probe_read_sleepable(&v58, sizeof(v58), (void *)ctx->args[1]+0) < 0)
-        return SECCOMP_RET_ERRNO | EINVAL;
+    	struct stat v58;
+    	if (bpf_probe_read_sleepable(&v58, sizeof(v58), (void *)ctx->args[1]+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
     //arg intptr intptr 0xf8ffa0 8
     //arg intptr intptr 0xf8ffa0 8
     //arg int32 int32 0xf8ffa0 4
@@ -75,7 +129,7 @@ int __always_inline filter_fstat(struct seccomp_data *ctx) {
     //arg const[0, const] const 0xf8ffa0 4
     //arg const[0, const] const 0xf8ffa0 4
 
-        info.id = 1;
+    	info.id = 1;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
