+++ complete_filter/ioctlKBASEIOCTLCSGETGLBIFACE.c
@@ -65,26 +65,62 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_cs_get_glb_iface {
+    uint32_t max_group_num; //int32
+    uint32_t max_total_stream_num; //int32
+    uint64_t groups_ptr; //ptr64
+    uint64_t streams_ptr; //ptr64
+    uint32_t glb_version; //int32
+    uint32_t features; //int32
+    uint32_t group_num; //int32
+    uint32_t prfcnt_size; //int32
+    uint32_t total_stream_num; //int32
+    uint32_t instr_featurs; //int32
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
 ({                                                             \
     char ____fmt[] = fmt;                                      \
     bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
-})*/
+})
+*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_CS_GET_GLB_IFACE(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0xc0188033 && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[inout, kbase_ioctl_cs_get_glb_iface] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_cs_get_glb_iface kbase_ioctl_cs_get_glb_iface 0xf8ffa0 24
+    	struct kbase_ioctl_cs_get_glb_iface v60;
+    	if (bpf_probe_read_sleepable(&v60, sizeof(v60), (void *)ctx->args[2]+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg ptr[out, int64] ptr64 0xf8ffa0 8
+    //arg int64 int64 0xf8ffa0 8
+    	uint64_t v61;
+    	if (bpf_probe_read_sleepable(&v61, sizeof(v61), (void *)v60.groups_ptr+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
+    //arg ptr[out, int64] ptr64 0xf8ffa0 8
+    //arg int64 int64 0xf8ffa0 8
+    	uint64_t v62;
+    	if (bpf_probe_read_sleepable(&v62, sizeof(v62), (void *)v60.streams_ptr+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
 
-        info.id = 31;
+        info.id = 3;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +128,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_CS_GET_GLB_IFACE reject\n");
     }
     return ret;
 }
