+++ complete_filter/ioctlKBASEIOCTLCSQUEUETERMINATE.c
@@ -65,26 +65,35 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_cs_queue_terminate {
+    uint64_t buffer_gpu_addr; //gpu_addr
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
 ({                                                             \
     char ____fmt[] = fmt;                                      \
     bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
-})*/
+})
+*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_CS_QUEUE_TERMINATE(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0x40088029 && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[in, kbase_ioctl_cs_queue_terminate] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_cs_queue_terminate kbase_ioctl_cs_queue_terminate 0xf8ffa0 8
+    	struct kbase_ioctl_cs_queue_terminate v85;
+    	if (bpf_probe_read_sleepable(&v85, sizeof(v85), (void *)ctx->args[2]+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
 
-        info.id = 31;
+        info.id = 43;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +101,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_CS_QUEUE_TERMINATE reject\n");
     }
     return ret;
 }
