+++ complete_filter/ioctlKBASEIOCTLMEMEXECINIT.c
@@ -65,6 +65,10 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_mem_exec_init {
+    uint64_t va_pages; //int64
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
@@ -74,17 +78,22 @@ DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 })*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_MEM_EXEC_INIT(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0x40088026 && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[in, kbase_ioctl_mem_exec_init] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_mem_exec_init kbase_ioctl_mem_exec_init 0xf8ffa0 8
+    	struct kbase_ioctl_mem_exec_init v63;
+    	if (bpf_probe_read_sleepable(&v63, sizeof(v63), (void *)ctx->args[2]+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
+    //arg int64 int64 0xf8ffa0 8
 
-        info.id = 31;
+        info.id = 4;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +101,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_MEM_EXEC_INIT reject\n");
     }
     return ret;
 }
