+++ complete_filter/ioctlKBASEIOCTLREADUSERPAGE.c
@@ -65,6 +65,13 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_read_user_page {
+    uint32_t offset; //user_offsets
+    uint32_t padding; //const
+    uint32_t val_lo; //int32
+    uint32_t val_hi; //int32
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
@@ -74,17 +81,25 @@ DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 })*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_READ_USER_PAGE(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0xc008803c && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[inout, kbase_ioctl_read_user_page] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_read_user_page kbase_ioctl_read_user_page 0xf8ffa0 8
+    	struct kbase_ioctl_read_user_page v77;
+    	if (bpf_probe_read_sleepable(&v77, sizeof(v77), (void *)ctx->args[2]+0) < 0)
+        	return SECCOMP_RET_ERRNO | EINVAL;
+    //arg const[0, user_offsets] user_offsets 0xf8ffa0 4
+    //arg const[0, const] const 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
+    //arg int32 int32 0xf8ffa0 4
 
-        info.id = 31;
+        info.id = 27;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +107,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_READ_USER_PAGE reject\n");
     }
     return ret;
 }
