+++ complete_filter/ioctlKBASEIOCTLSETFLAGS.c
@@ -65,6 +65,10 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_set_flags {
+    uint32_t create_flags; //basep_context_create_kernel_flags
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
@@ -74,17 +78,22 @@ DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 })*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_SET_FLAGS(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0x40048001 && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[in, kbase_ioctl_set_flags] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_set_flags kbase_ioctl_set_flags 0xf8ffa0 4
+    	struct kbase_ioctl_set_flags v59;
+    	if (bpf_probe_read_sleepable(&v59, sizeof(v59), (void *)ctx->args[2]+0) < 0)
+       		return SECCOMP_RET_ERRNO | EINVAL;
+    //arg basep_context_create_kernel_flags basep_context_create_kernel_flags 0xf8ffa0 4
 
-        info.id = 31;
+        info.id = 2;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +101,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_SET_FLAGS reject\n");
     }
     return ret;
 }
