+++ complete_filter/ioctlKBASEIOCTLSTICKYRESOURCEMAP.c
@@ -65,6 +65,11 @@ struct syscall_info {
     uint64_t fd;
 };
 
+struct kbase_ioctl_sticky_resource_map {
+    uint64_t count; //len
+    uint64_t address; //ptr64
+};
+
 DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 /*
 #define bpf_printk(fmt, ...)                                   \
@@ -74,17 +79,39 @@ DEFINE_BPF_MAP_N(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
 })*/
 
 SEC("seccomp")
-int __always_inline filter_read(struct seccomp_data *ctx) {
+int __always_inline filter_ioctl_KBASE_IOCTL_STICKY_RESOURCE_MAP(struct seccomp_data *ctx) {
     int ret = SECCOMP_RET_ALLOW;
     char dev [] = "/dev/bifrost";
 
-    if (ctx->nr == 63 && bpf_check_fd(dev, ctx->args[0])) {
+    if (ctx->nr == 29 && ctx->args[1] == 0x4010801d && bpf_check_fd(dev, ctx->args[0])) {
         struct syscall_info info = {};
         info.fd = ctx->args[0];
 
-    //arg ptr[out, buffer] ptr 0xf8ffa0 8
+    //arg ptr[in, kbase_ioctl_sticky_resource_map] ptr 0xf8ffa0 8
+    //arg kbase_ioctl_sticky_resource_map kbase_ioctl_sticky_resource_map 0xf8ffa0 16
+    struct kbase_ioctl_sticky_resource_map v75;
+    if (bpf_probe_read_sleepable(&v75, sizeof(v75), (void *)ctx->args[2]+0) < 0)
+        return SECCOMP_RET_ERRNO | EINVAL;
+    //arg len len 0xf8ffa0 8
+    if (v75.count != 0x1) {
+        ret = SECCOMP_RET_ERRNO | EINVAL;
+    }
+    //arg ptr[in, array[int64]] ptr64 0xf8ffa0 8
+    //arg array[int64] array 0xf8ffa0 varlen
+    //Unhandled array v76;
+    uint64_t v76;
+    size_t array_v76_offset = 0;
+    size_t array_v76_end = v75.count * sizeof(v76);
+    if (array_v76_offset + sizeof(v76) >= array_v76_end) {
+        goto array_v76_end;
+    }
+    if (bpf_probe_read_sleepable(&v76, sizeof(v76), (void *)v75.address+array_v76_offset) < 0)
+        return SECCOMP_RET_ERRNO | EINVAL;
 
-        info.id = 31;
+    array_v76_offset += sizeof(v76);
+array_v76_end:
+
+        info.id = 26;
 
         if (ret == SECCOMP_RET_ALLOW) {
             uint64_t pid_tgid = bpf_get_current_pid_tgid();
@@ -92,7 +119,7 @@ int __always_inline filter_read(struct seccomp_data *ctx) {
         }
     }
     if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("read reject\n");
+        bpf_printk("ioctl_KBASE_IOCTL_STICKY_RESOURCE_MAP reject\n");
     }
     return ret;
 }
