--- complete_filter/ioctl_KBASE_IOCTL_CS_QUEUE_REGISTER.c
@@ -1,64 +1,0 @@
-#include <linux/seccomp.h>
-#include <linux/bpf.h>
-#include <linux/unistd.h>
-#include <linux/ptrace.h>
-#include <bpf_helpers.h>
-#include <linux/errno.h>
-#include <sys/types.h>
-#include "tracer_id.h"
-
-struct syscall_info {
-    uint8_t id;
-    uint64_t fd;
-};
-
-struct base_ioctl_cs_queue_register {
-    uint64_t buffer_gpu_addr; //gpu_addr
-    uint32_t buffer_size; //int32
-    uint8_t priority; //int8
-    char padding[6]; //array
-    char pad1[5]; //pad
-};
-
-DEFINE_BPF_MAP(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
-#define bpf_printk(fmt, ...)                                   \
-({                                                             \
-    char ____fmt[] = fmt;                                      \
-    bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
-})
-
-SEC("seccomp")
-int __always_inline filter_ioctl_KBASE_IOCTL_CS_QUEUE_REGISTER(struct seccomp_data *ctx) {
-    int ret = SECCOMP_RET_ALLOW;
-    char dev [] = "/dev/bifrost";
-
-    if (ctx->nr == 29 && ctx->args[1] == 0x40108024 && bpf_check_fd(dev, ctx->args[0])) {
-        struct syscall_info info = {};
-        info.fd = ctx->args[0];
-
-    //arg ptr[in, base_ioctl_cs_queue_register] ptr 0xf8ffa0 8
-    //arg base_ioctl_cs_queue_register base_ioctl_cs_queue_register 0xf8ffa0 24
-    struct base_ioctl_cs_queue_register v69;
-    if (bpf_probe_read_sleepable(&v69, sizeof(v69), (void *)ctx->args[2]+0) < 0)
-        return SECCOMP_RET_ERRNO | EINVAL;
-    //arg int32 int32 0xf8ffa0 4
-    if (v69.buffer_size != 0x10000
-        && v69.buffer_size != 0x8000) {
-        ret = SECCOMP_RET_ERRNO | EINVAL;
-    }
-    //arg int8 int8 0xf8ffa0 1
-
-        info.id = 15;
-
-        if (ret == SECCOMP_RET_ALLOW) {
-            uint64_t pid_tgid = bpf_get_current_pid_tgid();
-            bpf_syscall_info_map_update_elem(&pid_tgid, &info, BPF_ANY);
-        }
-    }
-    if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("ioctl_KBASE_IOCTL_CS_QUEUE_REGISTER reject\n");
-    }
-    return ret;
-}
-
-char _license[] SEC("license") = "GPL";
