--- complete_filter/ioctl_KBASE_IOCTL_MEM_ALLOC_EX.c
@@ -1,103 +1,0 @@
-#include <linux/seccomp.h>
-#include <linux/bpf.h>
-#include <linux/unistd.h>
-#include <linux/ptrace.h>
-#include <bpf_helpers.h>
-#include <linux/errno.h>
-#include <sys/types.h>
-#include "tracer_id.h"
-
-struct syscall_info {
-    uint8_t id;
-    uint64_t fd;
-};
-
-struct kbase_ioctl_mem_alloc_ex {
-    uint64_t va_pages; //int64
-    uint64_t commit_pages; //int64
-    uint64_t extension; //int64
-    uint64_t flags; //base_mem_alloc_flags
-    uint64_t fixed_address; //int64
-    char extra[24]; //array
-    uint64_t out_flags; //int64
-    uint64_t gpu_va; //gpu_addr
-};
-
-DEFINE_BPF_MAP(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
-#define bpf_printk(fmt, ...)                                   \
-({                                                             \
-    char ____fmt[] = fmt;                                      \
-    bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
-})
-
-SEC("seccomp")
-int __always_inline filter_ioctl_KBASE_IOCTL_MEM_ALLOC_EX(struct seccomp_data *ctx) {
-    int ret = SECCOMP_RET_ALLOW;
-    char dev [] = "/dev/bifrost";
-
-    if (ctx->nr == 29 && ctx->args[1] == 0xc040803b && bpf_check_fd(dev, ctx->args[0])) {
-        struct syscall_info info = {};
-        info.fd = ctx->args[0];
-
-    //arg ptr[inout, kbase_ioctl_mem_alloc_ex] ptr 0xf8ffa0 8
-    //arg kbase_ioctl_mem_alloc_ex kbase_ioctl_mem_alloc_ex 0xf8ffa0 64
-    struct kbase_ioctl_mem_alloc_ex v65;
-    if (bpf_probe_read_sleepable(&v65, sizeof(v65), (void *)ctx->args[2]+0) < 0)
-        return SECCOMP_RET_ERRNO | EINVAL;
-    //arg int64 int64 0xf8ffa0 8
-    //arg int64 int64 0xf8ffa0 8
-    //arg int64 int64 0xf8ffa0 8
-    //arg base_mem_alloc_flags base_mem_alloc_flags 0xf8ffa0 8
-    //arg int64 int64 0xf8ffa0 8
-    //arg int64 int64 0xf8ffa0 8
-
-        if (v65.flags == 23) {
-            info.id = 6;
-        } else if (v65.flags == 14351) {
-            info.id = 9;
-        } else if (v65.flags == 8207) {
-            info.id = 10;
-        } else if (v65.flags == 524303) {
-            info.id = 11;
-        } else if (v65.flags == 2105359) {
-            info.id = 14;
-        } else if (v65.flags == 533519) {
-            info.id = 19;
-        } else if (v65.flags == 10255) {
-            info.id = 20;
-        } else if (v65.flags == 4204559) {
-            info.id = 21;
-        } else if (v65.flags == 12303) {
-            info.id = 22;
-        } else if (v65.flags == 14343) {
-            info.id = 23;
-        } else if (v65.flags == 8204) {
-            info.id = 29;
-        } else if (v65.flags == 8199) {
-            info.id = 32;
-        } else if (v65.flags == 4206607) {
-            info.id = 34;
-        } else if (v65.flags == 4208655) {
-            info.id = 35;
-        } else if (v65.flags == 4203023) {
-            info.id = 36;
-        } else if (v65.flags == 4202511) {
-            info.id = 38;
-        } else if (v65.flags == 4205069) {
-            info.id = 46;
-        } else if (v65.flags == 8268) {
-            info.id = 49;
-        }
-
-        if (ret == SECCOMP_RET_ALLOW) {
-            uint64_t pid_tgid = bpf_get_current_pid_tgid();
-            bpf_syscall_info_map_update_elem(&pid_tgid, &info, BPF_ANY);
-        }
-    }
-    if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("ioctl_KBASE_IOCTL_MEM_ALLOC_EX reject\n");
-    }
-    return ret;
-}
-
-char _license[] SEC("license") = "GPL";
