--- complete_filter/ioctl_KBASE_IOCTL_READ_USER_PAGE.c
@@ -1,61 +1,0 @@
-#include <linux/seccomp.h>
-#include <linux/bpf.h>
-#include <linux/unistd.h>
-#include <linux/ptrace.h>
-#include <bpf_helpers.h>
-#include <linux/errno.h>
-#include <sys/types.h>
-#include "tracer_id.h"
-
-struct syscall_info {
-    uint8_t id;
-    uint64_t fd;
-};
-
-struct kbase_ioctl_read_user_page {
-    uint32_t offset; //user_offsets
-    uint32_t padding; //const
-    uint32_t val_lo; //int32
-    uint32_t val_hi; //int32
-};
-
-DEFINE_BPF_MAP(syscall_info_map, HASH, uint64_t, struct syscall_info, 512);
-#define bpf_printk(fmt, ...)                                   \
-({                                                             \
-    char ____fmt[] = fmt;                                      \
-    bpf_trace_printk(____fmt, sizeof(____fmt), ##__VA_ARGS__); \
-})
-
-SEC("seccomp")
-int __always_inline filter_ioctl_KBASE_IOCTL_READ_USER_PAGE(struct seccomp_data *ctx) {
-    int ret = SECCOMP_RET_ALLOW;
-    char dev [] = "/dev/bifrost";
-
-    if (ctx->nr == 29 && ctx->args[1] == 0xc008803c && bpf_check_fd(dev, ctx->args[0])) {
-        struct syscall_info info = {};
-        info.fd = ctx->args[0];
-
-    //arg ptr[inout, kbase_ioctl_read_user_page] ptr 0xf8ffa0 8
-    //arg kbase_ioctl_read_user_page kbase_ioctl_read_user_page 0xf8ffa0 8
-    struct kbase_ioctl_read_user_page v77;
-    if (bpf_probe_read_sleepable(&v77, sizeof(v77), (void *)ctx->args[2]+0) < 0)
-        return SECCOMP_RET_ERRNO | EINVAL;
-    //arg const[0, user_offsets] user_offsets 0xf8ffa0 4
-    //arg const[0, const] const 0xf8ffa0 4
-    //arg int32 int32 0xf8ffa0 4
-    //arg int32 int32 0xf8ffa0 4
-
-        info.id = 27;
-
-        if (ret == SECCOMP_RET_ALLOW) {
-            uint64_t pid_tgid = bpf_get_current_pid_tgid();
-            bpf_syscall_info_map_update_elem(&pid_tgid, &info, BPF_ANY);
-        }
-    }
-    if (ret != SECCOMP_RET_ALLOW) {;
-        bpf_printk("ioctl_KBASE_IOCTL_READ_USER_PAGE reject\n");
-    }
-    return ret;
-}
-
-char _license[] SEC("license") = "GPL";
